% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number_mfe.R
\name{format_num}
\alias{format_num}
\title{A Function to Format Numbers}
\usage{
format_num(number_value, accuracy = 0.01)
}
\arguments{
\item{number_value}{the number to modify}
}
\description{
Consistenly formatting of numbers.
}
\examples{
format_num(1520000)
> "1,520,000.00"
format_num(0.00008)
> "0.00"
}
\keyword{number-formatting}
