% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaya.R
\docType{data}
\name{kaya}
\alias{kaya}
\title{Kaya dataset}
\format{
An object of class \code{data.frame} with 5 rows and 30 columns.
}
\usage{
data(kaya)
}
\description{
A brief description
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{iplotCurves(phe, times)}
}
\keyword{datasets}
