% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_macrons.R
\name{mac}
\alias{mac}
\title{Format macrons}
\usage{
mac(x)
}
\arguments{
\item{x}{the vowel that we want to put a macron on}
}
\description{
Function to apply MfE branded colours
}
\examples{
mac()

}
\keyword{macron-formatting}
