% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maori_land.R
\docType{data}
\name{maori_land}
\alias{maori_land}
\title{maori land dataset}
\format{
An object of class \code{data.frame} with 792 rows and 4 columns.
}
\usage{
data(maori_land)
}
\description{
A brief description
}
\examples{
data(grav)
times <- attr(grav, "time")
phe <- grav$pheno
\donttest{iplotCurves(phe, times)}
}
\keyword{datasets}
