% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfe_colours.r
\docType{data}
\name{mfe_colours}
\alias{mfe_colours}
\title{MfE set upa list of MfE colours}
\format{
An object of class \code{list} of length 7.
}
\usage{
mfe_colours
}
\description{
MfE set upa list of MfE colours
}
\examples{
mfe_colors()
}
\keyword{ggplot-colour-formatting}
