% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfe_pal.r
\name{mfe_pal}
\alias{mfe_pal}
\title{MfE set up mfe_pal()}
\usage{
mfe_pal(primary = "dark_green", other = "orange", direction = 1)
}
\description{
MfE set up mfe_pal()
}
\examples{
mfe_pal()
}
\keyword{ggplot-colour-formatting}
