% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_mfe.R
\name{scale_colour_mfe}
\alias{scale_colour_mfe}
\title{MfE discrete colours for ggplot2 plots}
\usage{
scale_colour_mfe(primary = "dark_blue", other = "orange", direction = 1, ...)
}
\arguments{
\item{primary}{defaults to dark_green}

\item{other}{defaults to orange}

\item{direction}{defaults to 1}
}
\description{
Function to apply MfE branded colours
}
\examples{
scale_colour_mfe()
}
\keyword{ggplot-colour-formatting}
