% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_mfe_d.R
\name{scale_fill_mfe}
\alias{scale_fill_mfe}
\title{MfE discrete filling for ggplot2 plots}
\usage{
scale_fill_mfe(primary = "light_blue", other = "orange", direction = 1, ...)
}
\arguments{
\item{primary}{defaults to dark_green}

\item{other}{defaults to orange}

\item{direction}{defaults to 1}
}
\description{
Function to apply MfE branded colours
}
\examples{
scale_fill_mfe()
}
\keyword{ggplot-colour-formatting}
