% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_theme_inventory.R
\name{theme_inventory}
\alias{theme_inventory}
\title{A Function by Dan Lawrence}
\usage{
theme_inventory(
  font_family = "Helvetica",
  font_size_title = 11,
  font_size_body = 10
)
}
\arguments{
\item{font_family}{the value that you are trying to modify}
}
\description{
This function allows you to express your love of cats.
}
\examples{
theme_inventory()
}
\keyword{ggplot-formatting}
