% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_format_number.R
\name{format_num}
\alias{format_num}
\title{A Function by Dan Lawrence}
\usage{
format_num(number_value, accuracy = 0.01)
}
\arguments{
\item{number_value}{the value that you are trying to modify}
}
\description{
This function allows you to express your love of cats.
}
\examples{
format_num()
}
\keyword{number-formatting}
