% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_theme_inventory.R
\name{theme_inventory}
\alias{theme_inventory}
\title{MfE theme for ggplot2 plots}
\usage{
theme_inventory(
  font_family = "Helvetica",
  font_size_title = 11,
  font_size_body = 10
)
}
\arguments{
\item{font_family}{defaults to Helvetica}

\item{font_size_title}{defaults to 11}

\item{font_size_body}{defaults to 10}
}
\description{
Function to apply MfE theme (for now, copied form simplevis)
}
\examples{
ggplot(msleep, aes( x = sleep_total , y = reorder(name, -sleep_total ))) +
geom_point()+
theme_inventory()
}
\keyword{ggplot-formatting}
